+++ src/websockets/datastructures.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Iterable, Iterator, Mapping, MutableMapping
-from typing import Any, Protocol, Union
+from typing import Any, Protocol
 
 
 __all__ = [
@@ -171,13 +171,9 @@ class SupportsKeysAndGetItem(Protocol):  # pragma: no cover
     def __getitem__(self, key: str) -> str: ...
 
 
-# Change to Headers | Mapping[str, str] | ... when dropping Python < 3.10.
-HeadersLike = Union[
-    Headers,
-    Mapping[str, str],
-    Iterable[tuple[str, str]],
-    SupportsKeysAndGetItem,
-]
+HeadersLike = (
+    Headers | Mapping[str, str] | Iterable[tuple[str, str]] | SupportsKeysAndGetItem
+)
 """
 Types accepted where :class:`Headers` is expected.