+++ src/websockets/legacy/client.py
@@ -607,16 +607,14 @@ class Connect:
                     self.logger.info(
                         "connect failed; reconnecting in %.1f seconds: %s",
                         initial_delay,
-                        # Remove first argument when dropping Python 3.9.
-                        traceback.format_exception_only(type(exc), exc)[0].strip(),
+                        traceback.format_exception_only(exc)[0].strip(),
                     )
                     await asyncio.sleep(initial_delay)
                 else:
                     self.logger.info(
                         "connect failed again; retrying in %d seconds: %s",
                         int(backoff_delay),
-                        # Remove first argument when dropping Python 3.9.
-                        traceback.format_exception_only(type(exc), exc)[0].strip(),
+                        traceback.format_exception_only(exc)[0].strip(),
                     )
                     await asyncio.sleep(int(backoff_delay))
                 # Increase delay with truncated exponential backoff.
@@ -673,7 +671,7 @@ class Connect:
             else:
                 raise SecurityError("too many redirects")
 
-    # ... = yield from connect(...) - remove when dropping Python < 3.10
+    # ... = yield from connect(...) - remove when dropping Python < 3.11
 
     __iter__ = __await__